+++ src/components/course/common/ErrorMessage.tsx
@@ -1,4 +1,4 @@
-import Button from '../commons/button/Button'
+import Button from '@src/components/commons/button/Button'
 import Icon from '@src/components/commons/Icon'
 import { BUTTON_TEXT, ERROR_UI, ERROR_ICONS } from '@src/constants/ui'
 import type { LucideIcon } from 'lucide-react'
