+++ src/components/course/course/CourseStats.tsx
