+++ src/components/course/sections/ReviewSection.tsx
@@ -1,4 +1,4 @@
-import StarRating from './StarRating'
+import StarRating from '../ui/StarRating'
 import type { Review } from '@src/mock/reviewData'
 
 interface ReviewItemProps {
