+++ src/components/course/ui/NavigationButton.tsx
