+++ src/components/course/ui/SearchInput.tsx
