+++ src/components/course/ui/SectionHeader.tsx
