+++ src/components/course/ui/StarRating.tsx
